+++ sources/android/telecom/Logging/TimedEvent.java
@@ -4,6 +4,7 @@ import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 
+/* loaded from: classes3.dex */
 public abstract class TimedEvent<T> {
     public abstract T getKey();
 
@@ -15,14 +16,14 @@ public abstract class TimedEvent<T> {
         for (TimedEvent<T> entry : events) {
             if (counts.containsKey(entry.getKey())) {
                 counts.put(entry.getKey(), Integer.valueOf(counts.get(entry.getKey()).intValue() + 1));
-                result.put(entry.getKey(), Double.valueOf(result.get(entry.getKey()).doubleValue() + ((double) entry.getTime())));
+                result.put(entry.getKey(), Double.valueOf(result.get(entry.getKey()).doubleValue() + entry.getTime()));
             } else {
                 counts.put(entry.getKey(), 1);
-                result.put(entry.getKey(), Double.valueOf((double) entry.getTime()));
+                result.put(entry.getKey(), Double.valueOf(entry.getTime()));
             }
         }
         for (Map.Entry<T, Double> entry2 : result.entrySet()) {
-            result.put(entry2.getKey(), Double.valueOf(entry2.getValue().doubleValue() / ((double) counts.get(entry2.getKey()).intValue())));
+            result.put(entry2.getKey(), Double.valueOf(entry2.getValue().doubleValue() / counts.get(entry2.getKey()).intValue()));
         }
         return result;
     }
